Very rough work in progress
The approach I've started here is not a great one. It relies on recognizing
the strings which need to be colorized rather than having the data broken out.
In other words, the list of list of object is really hindering this being written the
way it should be. Whoever works on this next should strongly consider refactoring.